Fixed bugs in logic and added additional measurements in debug mode

+++ PrisonersProblem/Models/Classes/ProblemStatistic.cs
@@ -5,7 +5,7 @@ public class ProblemStatistic : IProblemStatistic
     public int WinNumber { get; set; } = 0;
     public int LoseNumber { get; set; } = 0;
     public int TotalNumber { get => WinNumber + LoseNumber; }
-    public int WinsPercentage { get => WinNumber / TotalNumber * 100; }
+    public double WinsPercentage { get => (double)WinNumber / (double)TotalNumber * 100.0; }
 
     public new string ToString()
     {
