+++ PrisonersProblem/Models/Interfaces/IProblemStatistic.cs
@@ -5,7 +5,7 @@ public interface IProblemStatistic
     public int WinNumber { get; set; }
     public int LoseNumber { get; set; }
     public int TotalNumber { get => WinNumber + LoseNumber; }
-    public int WinsPercentage { get => WinNumber / TotalNumber * 100; }
+    public double WinsPercentage { get => WinNumber / TotalNumber * 100; }
 
     public string ToString();
 }
